# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Learning with LLMs API
  version: 1.0.0
  description: API for the Learning with LLMs educational chatbot application

servers:
  - url: http://localhost:5001
    description: Local development server

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns the health status of the API server
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true

  /chat:
    post:
      summary: Send a message to the AI tutor
      description: Send a message to the AI tutor and receive a response based on the learning stage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Successful response from the AI tutor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Message is required"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "OpenAI API request failed"

  /session/summary:
    post:
      summary: Generate a summary of a study session
      description: Generate a concise summary of a study session based on chat history
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionSummaryRequest'
      responses:
        '200':
          description: Successful summary generation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionSummaryResponse'
        '400':
          description: Bad request - missing sessionId or no messages found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "sessionId is required"
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Session not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to generate session summary"

components:
  schemas:
    ChatRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: The user's message to the tutor
          example: "Explain gradients simply"
        stage:
          type: integer
          description: Learning stage (1-4)
          minimum: 1
          maximum: 4
          default: 1
          example: 1
        sessionId:
          type: string
          description: Session ID for conversation continuity
          example: "64f8a1b2c3d4e5f6a7b8c9d0"

    ChatResponse:
      type: object
      required:
        - sessionId
        - reply
      properties:
        sessionId:
          type: string
          description: Session ID for conversation continuity
          example: "64f8a1b2c3d4e5f6a7b8c9d0"
        reply:
          type: string
          description: The AI tutor's response
          example: "Gradients are a fundamental concept in calculus that measure how steep a function is at any point..."

    SessionSummaryRequest:
      type: object
      required:
        - sessionId
      properties:
        sessionId:
          type: string
          description: The session ID to summarize
          example: "64f8a1b2c3d4e5f6a7b8c9d0"

    SessionSummaryResponse:
      type: object
      required:
        - sessionId
        - summary
      properties:
        sessionId:
          type: string
          description: The session ID that was summarized
          example: "64f8a1b2c3d4e5f6a7b8c9d0"
        summary:
          type: string
          description: The generated summary of the study session
          example: "• Key concept: Derivatives measure rate of change\n• Application: Used in optimization problems\n• Next steps: Practice with more complex functions"